// client/src/app/classes/svg-properties.ts
export enum SVGProperties {
  centerX = 'cx',
  centerY = 'cy',
  color = 'stroke',
  fill = 'fill',
  fillOpacity = 'fill-opacity',
  height = 'height',
  hyperlink = 'a',
  joint = 'stroke-linejoin',
  lineColor = 'stroke',
  opacity = 'opacity',
  points = 'points',
  polyLine = 'polyline',
  position = 'viewBox',
  R = 'r',
  radius = 'r',
  radiusX = 'rx',
  radiusY = 'ry',
  extremities = 'stroke-linecap',
  colorOpacity = 'stroke-opacity',
  title = 'title',
  thickness = 'stroke-width',
  width = 'width',
  X = 'x',
  Y = 'y'
}